pyproject.toml
[tool.poetry]
name = "alisa-gpt-assistant"
version = "0.1.0"
description = "OpenAI GPT Assistant integration with Yandex Alisa via dialogs.yandex.ru"
authors = ["Ivan Reshetnikov <resivalex@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.3.7"
black = {extras = ["d"], version = "^23.11.0"}
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
